refactor(renderMovieHome): 영화 홈화면 그리는 함수 분리
